fix: all keyboards lookup performance
Addresses issues with performance for looking up all keyboards, e.g. as
used by KeymanWeb:
https://api.keyman.com/cloud/4.0/keyboards?jsonp=keyman.register&languageidtype=bcp47&version=12.0.103&timerid=6

This would result in about 630 separate queries, one for the list of
keyboards, and one for each keyboard returned to get the list of
languages associated with the keyboards. The fix reduces this to
two queries: one for the keyboards, and one for the languages.

This appears to be correlated to a massive increase in data egress on
api.keyman.com MySQL database instance; while the queries themselves
amounted to about 1.3mb of data, each call would result in 80mb of
outbound traffic. This may be due to some configuration change on the
database as it has only arisen recently.

In any case, this change results in a significant performance
improvement and should mitigate the data usage immediately.

Note that the api.keyman.com endpoint is proxied by CloudFlare and you
may want to test against keymanapi.azurewebsites.net to observe the
performance before and after.