fix: Improve performance of legacy/languages
`getLanguages()` was also very slow in legacy30 and legacy40, for
opposite reason of the last commit: in this case everything was being
returned in one monolithic query which duplicated data heavy fields,
resulting in a 70MB query result, which was transformed finally into
a 270KB JSON result.

Splitting this query into 2 queries allowed for normalization of this
data; this query was also suffering from the same issue in legacy30.php
as was fixed in legacy40.php in #24, which creates a massive performance
hit (namely spurious Notices when looking up missing fonts/examples).