fix: id search returning all results for Unicode
Because the keyboard_id search was using a LIKE comparison with a
Latin1_General collation (which is appropriate given keyboard ids are
a Latin1 style string), non-Latin characters such as ት were causing the
search to return every possible keyboard (1098 of them!), which was
probably not *quite* what we wanted. Changed to strip all non-id-
relevant characters from the search -- no decomposition etc -- so that
only exact character matches (from start of string) will work.

Also tweaked slightly so that if a zero-length id string is searched, no
results will be returned.