feat: major version upgrades (Windows)
Fixes keymanapp/keyman#1635.

This implements the major version upgrade functionality for Keyman on
Windows. There are quite a few nuances to take into account:

1. We offer updates for alpha to alpha versions -- they'll never go to
   beta or stable.
2. Beta versions will be updated to stable, as the beta tier goes stale
   between major releases.
3. Stable only ever upgrades to stable.
4. We rollout stable major releases gradually to ensure that we don't
   have hundreds of thousands of users updating at once.
5. Keyman Desktop 10.0 does not currently do major version upgrades,
   because of a limitation in the installer which means that user
   keyboards are not migrated (see keymanapp/keyman#3865).
6. This does not support upgrading versions earlier than 10.0 at all.
7. This upgrades Keyman Desktop 11.0 onwards, and Keyman Developer
   11.0-13.0. Keyman Developer 14.0 upgrades are awaiting code updates
   in the keymanapp/keyman repo, and will likely require a new PR on
   api.keyman.com for full support also (keymanapp/keyman#3846).
8. If a user manually checks for updates, they'll receive the major
   version upgrade offer immediately, while the automated weekly check
   still works in the rollout schedule. (Corresponding keymanapp/keyman
   PR to come).

Whew. Online updates are supposed to be easy, right?

I have not yet written tests for the various permutations of manual vs
automatic update checks and tiers. I am currently pretending I don't
need to because it's not a fun job... and will probably require some
additional refactoring in ordre for ReleaseSchedule to be side-effect
free when used from WindowsUpdateCheck etc.