fix: cached keyboard data not refreshing
The keyboard build web job needs to ignore the cache in order to always
get the latest keyboard and model data.

In theory we could make this signal only for the keyboard and model data
and not for the other standards data, but I think that this will be okay
anyway, as we don't run more than a handful of builds a day.