fix: improve weight for searches
Fixes #147.

The search results for 'german' would return a set of weights for
various language codes. 'de' had 5 weights and 'pdt-Latn' had 4, but
both had a match on the name 'German' because 'pdt-Latn' has an
alternate language name of 'German'. This meant that the actual result
returned was indeterminate -- they both had the same weight.

Summing the weights for each language tag match brought 'de' higher
in the match results (weight 60 vs 30).