fix: avoid missing packages with spaces or periods in id
Relates to keymanapp/keyman#4886.

If a package id has a period or a space (which can happen with legacy
packages), then the online update API would not find it, because of a
PHP misfeature which transforms these characters in input query
parameter names to underscore (really!). This seems like it is
potentially impossible to work around for POST requests (`php://input`
does not work for `multipart/form-data` duh), so I have added an
alternative parameter layout instead. Obviously, this means that Keyman
for Windows requires a corresponding update to use the new parameter
layout.