fix: Download statistics in local database
Fixes keymanapp/keyman.com#297.

Corresponding PRs for keyman.com and help.keyman.com coming shortly.

Instead of reading keyboard download statistics from Google Analytics
periodically, we now use the local sql database to store them. The stats
are very basic, just downloads per day per keyboard for now.

The statistics are stored in a separate schema, `kstats`, which will not
be overwritten when the database is upgraded. This means that any
changes must be managed carefully with database transitions rather than
cut-and-shut.

The new endpoint /increment-download/<keyboard_id> will increment the
download counter for the specified keyboard for the current day. The
incremented result is returned and can also be seen instantly by
reloading the keyboard detail page. This is an improvement on the old
model where new stats would only be available after a few days.

The endpoint /increment-download/<keyboard_id> will not be documented as
a public API.