feat: .kps XML Schema
Defines a reasonable XML schema for .kps files. This has been tested
against keyboards repository, and it picked up a couple of minor errors
within it.

For now, not attempting to build this into Keyman Developer, but will
add an xmllint step to keyboards repo soon for .kps files.

Part of keymanapp/keyman#5464.