chore: remove exception handler for Keyman Desktop 10
We no longer do anything with crash reports from early versions of
Keyman, so there's no point in collecting the data. This also resolves
a crash when handling some reports.

This also removes the need to have the api_keyman_com_github_user and
api_keyman_com_github_repo variables defined, which will be cleaned up
as a future task.

Fixes: #249
Fixes: API-KEYMAN-COM-4H