fix: wait for full text indexing to complete when building database
We started to see situations where status checks for new PRs would fail,
for example see #290, with hard to parse errors about actual search
results mismatching expected results. After digging in, with a lot of
logging and debugging, I found that the issue was a race where the test
would run before full text indexing had completed. This change delays
the completion of the database build until the full text indexing has
finished.

The full description of my investigation is recorded in #290 for future
reference.