fix: select existing tag name in langtag canonicalization
Problem description in #312; this fix just uses the existing langtags
name rather than the name from the .keyboard_info, which could vary. The
name is used only as a display name, not a search name, so no searching
will be impacted. It is possible that a different name may be selected
as representative, but that is probably better because we should be
using langtags.json for language names in most contexts anyway.

Fixes: #312